move version.go
