+++ pkg/version/version.go
@@ -1,4 +1,4 @@
-package utils
+package version
 
 import (
 	"fmt"
