+++ pkg/version/version_test.go
@@ -1,4 +1,4 @@
-package utils
+package version
 
 import (
 	"testing"
